refactor(cda-main): Move `AppError` into separate `error` module.
